feat: embrace proxies
