+++ .github/workflows/npm-publish.yml
@@ -2,9 +2,9 @@ name: Node.js Package
 
 on:
   release:
-    types: [created]
+    types: [ created ]
   workflow_dispatch:
-    branches: [master]
+    branches: [ master ]
 
 jobs:
   build:
