+++ src/models/anime-parser.ts
@@ -1,6 +1,10 @@
-import { BaseParser, IAnimeInfo, ISource, IEpisodeServer } from '.';
+import { BaseParser, IAnimeInfo, ISource, IEpisodeServer, ProxyConfig } from '.';
 
 abstract class AnimeParser extends BaseParser {
+  constructor(baseUrl?: string, proxyConfig?: ProxyConfig) {
+    super(baseUrl, proxyConfig);
+  }
+
   /**
    * if the provider has dub and it's avialable seperatly from sub set this to `true`
    */
