+++ src/models/types.ts
@@ -411,7 +411,6 @@ export enum Genres {
   CARS = 'Cars',
   COMEDY = 'Comedy',
   DRAMA = 'Drama',
-  ECCHI = 'Ecchi',
   FANTASY = 'Fantasy',
   HORROR = 'Horror',
   MAHOU_SHOUJO = 'Mahou Shoujo',
